refactor: dividir responsabilidades rutas
