+++ backend/routes/auth_routes.py
@@ -1,0 +1,22 @@
+from fastapi import APIRouter,Depends, HTTPException, status
+from fastapi.security import OAuth2PasswordRequestForm
+from sqlalchemy.orm import Session
+from database.database import get_db
+from services.auth_service import AuthService
+from schemas.user_schemas import (
+    Token
+)
+
+router = APIRouter()
+
+@router.post("/token", response_model=Token)
+def login(form_data: OAuth2PasswordRequestForm = Depends(), db: Session = Depends(get_db)):
+    usuario = AuthService.autenticar_usuario(db, form_data.username, form_data.password)
+    if not usuario:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Usuario o contraseña incorrectos",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
+    access_token = AuthService.crear_token_acceso(data={"sub": usuario.nombre_usuario})
+    return {"access_token": access_token, "token_type": "bearer"}
